requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
neo4j==5.27.0
packaging==24.2
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.1.0
Werkzeug==3.1.3
wheel==0.44.0
